make output files for synthetic seismograms (Mineos and Axisem so far)
