+++ burnman/__init__.py
@@ -76,6 +76,7 @@ from . import minerals
 
 # central user tools
 from . import seismic
+from . import output_seismo
 from . import averaging_schemes
 from . import eos
